+++ application/ui/src/features/models/train-model/model-architectures-list/model-architecture/model-architecture.module.scss
@@ -18,7 +18,7 @@
 
 .modelArchitectureSelected {
     border: var(--spectrum-alias-border-size-thick) solid var(--energy-blue);
-    background-color: #0c1e29;
+    background-color: var(--selection-dark-blue);
 }
 
 .modelArchitectureName {
