+++ application/ui/src/features/models/train-model/model-architectures-list/model-architectures-list-layout/model-architectures-list-layout.module.scss
