+++ application/ui/src/features/models/train-model/model-architectures-list/recommended-model-architectures.component.tsx
@@ -1,0 +1,73 @@
+// Copyright (C) 2025 Intel Corporation
+// SPDX-License-Identifier: Apache-2.0
+
+import { Flex } from '@geti/ui';
+
+import type { ModelArchitecture as ModelArchitectureType } from '../../../../constants/shared-types';
+import { ModelArchitectureCard } from './model-architecture/model-architecture.component';
+import { ModelArchitecturesListLayout } from './model-architectures-list-layout/model-architectures-list-layout.component';
+
+interface RecommendedModelArchitectureProps {
+    activeModelArchitectureId: string | undefined;
+    modelArchitecture: ModelArchitectureType;
+    selectedModelArchitectureId: string | null;
+    onSelectedModelArchitectureIdChange: (modelArchitectureId: string | null) => void;
+}
+
+const RecommendedModelArchitecture = ({
+    activeModelArchitectureId,
+    modelArchitecture,
+    selectedModelArchitectureId,
+    onSelectedModelArchitectureIdChange,
+}: RecommendedModelArchitectureProps) => {
+    const isSelected = modelArchitecture.id === selectedModelArchitectureId;
+    const isActive = activeModelArchitectureId === modelArchitecture.id;
+
+    return (
+        <ModelArchitectureCard
+            modelArchitecture={modelArchitecture}
+            isSelected={isSelected}
+            onSelect={() => onSelectedModelArchitectureIdChange(modelArchitecture.id)}
+        >
+            <Flex width={'100%'} minWidth={0} direction={'column'} gap={'size-100'}>
+                {isActive && <ModelArchitectureCard.Active />}
+                <ModelArchitectureCard.Name />
+            </Flex>
+            <ModelArchitectureCard.Parameters />
+            <ModelArchitectureCard.Divider />
+            <ModelArchitectureCard.ExpandedDescription />
+        </ModelArchitectureCard>
+    );
+};
+
+interface RecommendedModelArchitectures {
+    activeModelArchitectureId: string | undefined;
+    modelArchitectures: ModelArchitectureType[];
+    selectedModelArchitectureId: string | null;
+    onSelectedModelArchitectureIdChange: (modelArchitectureId: string | null) => void;
+}
+
+export const RecommendedModelArchitectures = ({
+    activeModelArchitectureId,
+    modelArchitectures,
+    onSelectedModelArchitectureIdChange,
+    selectedModelArchitectureId,
+}: RecommendedModelArchitectures) => {
+    return (
+        <ModelArchitecturesListLayout
+            selectedModelArchitectureId={selectedModelArchitectureId}
+            onSelectedModelArchitectureIdChange={onSelectedModelArchitectureIdChange}
+            ariaLabel={'Recommended model architectures'}
+        >
+            {modelArchitectures.map((modelArchitecture) => (
+                <RecommendedModelArchitecture
+                    key={modelArchitecture.id}
+                    activeModelArchitectureId={activeModelArchitectureId}
+                    modelArchitecture={modelArchitecture}
+                    selectedModelArchitectureId={selectedModelArchitectureId}
+                    onSelectedModelArchitectureIdChange={onSelectedModelArchitectureIdChange}
+                />
+            ))}
+        </ModelArchitecturesListLayout>
+    );
+};
